// src/app/models/UserDto.ts
export interface UserDto{
  userId: string;
    email: string;
    avatar: string | null;
    firstName: string;
    lastName: string;
    dateOfBirth: string | null;
    location: string | null;
    status: string;
    website: string | null;
    about: string | null;
    skype: string | null;
    facebook: string | null;
    twitter: string | null;
    role: string;
}